pyproject.toml
[project]
name = "deepsad-mnist"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "timm>=1.0.25",
    "scikit-learn>=1.8.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "matplotlib>=3.10.8",
    "tensorboard>=2.19.0",
    "setuptools>=68.0,<72.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
